Here is a draft for your commit message based on the provided diffs:
Implementação do repositório SQL e adequação da interface Database

- Criação da classe `SQLDatabase` implementando a interface `Database` para conexão com MySQL via JDBC.
- Implementação inicial dos métodos `listaItensCardapio`, `totalItensCardapio` e `adicionaItemCardapio` no repositório SQL.
- Alteração do tipo de retorno do método `totalItensCardapio` de `int` para `long` na interface `Database` e atualização correspondente na classe `ItemCardapioInMemoryRepository`.

+++ src/main/java/mx/com/florinda/repositories/Database.java
@@ -15,7 +15,7 @@ public interface Database {
 
     boolean alteraPrecoItemCardapio(Long id, double novoPreco);
 
-    int totalItensCardapio();
+    long totalItensCardapio();
 
     void adicionaItemCardapio(ItemCardapio item);
