+++ src/main/java/mx/com/florinda/repositories/ItemCardapioInMemoryRepository.java
@@ -45,7 +45,7 @@ public class ItemCardapioInMemoryRepository implements Database {
     }
 
     @Override
-    public int totalItensCardapio() {
+    public long totalItensCardapio() {
         return itens.size();
     }
